Finish initial pass of pclean_lib rewrite
