+++ cxx/BUILD
@@ -65,6 +65,7 @@ cc_library(
         ":irm",
         ":observations",
         "//distributions:crp",
+        "//pclean:get_joint_relations",
         "//pclean:io",
         "//pclean:schema",
     ],
