+++ cxx/gendb_test.cc
@@ -79,7 +79,7 @@ void test_unincorporate_reference_helper(GenDB& gendb,
   auto unincorporated_items = gendb.unincorporate_reference(
       class_name, ref_field, class_item, from_cluster_only);
 
-  const auto& ref_indices = gendb.schema_helper.relation_reference_indices;
+  const auto& ref_indices = gendb.relation_reference_indices;
   for (const auto& [name, trel] : gendb.hirm->schema) {
     // Store the indices of the relation domains that refer to the class and
     // the reference class.
@@ -308,6 +308,379 @@ BOOST_AUTO_TEST_CASE(test_update_reference_items) {
       BOOST_TEST(!unincorporated_items.at(rel_name).contains(items));
     }
   }
+}
+
+BOOST_AUTO_TEST_CASE(test_domains_cache) {
+  std::mt19937 prng;
+  GenDB gendb(&prng, schema);
+
+  std::vector<std::string> expected_domains = {"School"};
+  BOOST_TEST(gendb.domains["School"] == expected_domains);
+
+  expected_domains = {"School", "Physician"};
+  BOOST_TEST(gendb.domains["Physician"] == expected_domains);
+
+  expected_domains = {"City"};
+  BOOST_TEST(gendb.domains["City"] == expected_domains);
+
+  expected_domains = {"City", "Practice"};
+  BOOST_TEST(gendb.domains["Practice"] == expected_domains);
+
+  expected_domains = {"City", "Practice", "School", "Physician", "Record"};
+  BOOST_TEST(gendb.domains["Record"] == expected_domains, tt::per_element());
+
+  auto& ref_indices = gendb.class_reference_indices;
+
+  // The Practice, Physician, and Record classes have reference fields, so they
+  // should be included in the reference field index map.
+  BOOST_TEST(ref_indices.size() == 3);
+
+  // For Physician and Practice, index 1 corresponds to the class itself, and
+  // index 0 corresponds to the reference class.
+  BOOST_TEST_REQUIRE(ref_indices.contains("Physician"));
+  BOOST_TEST(ref_indices.at("Physician").at(1).at("school") == 0);
+  BOOST_TEST(ref_indices.at("Practice").at(1).at("city") == 0);
+
+  // For Record, index 4 corresponds to the class itself, which points to
+  // physician (index 3) and location (index 1).
+  BOOST_TEST_REQUIRE(ref_indices.contains("Record"));
+  BOOST_TEST(ref_indices.at("Record").at(4).at("physician") == 3);
+  BOOST_TEST(ref_indices.at("Record").at(4).at("location") == 1);
+  BOOST_TEST(ref_indices.at("Record").at(3).at("school") == 2);
+  BOOST_TEST(ref_indices.at("Record").at(1).at("city") == 0);
+}
+
+BOOST_AUTO_TEST_CASE(test_domains_and_reference_cache_two_paths_same_source) {
+  std::stringstream ss(R"""(
+class City
+  name ~ string
+
+class Person
+  birth_city ~ City
+  home_city ~ City
+)""");
+  PCleanSchema schema;
+  [[maybe_unused]] bool ok = read_schema(ss, &schema);
+  assert(ok);
+  std::mt19937 prng;
+  GenDB gendb(&prng, schema);
+
+  std::vector<std::string> expected_domains = {"City", "City", "Person"};
+  BOOST_TEST(gendb.domains["Person"] == expected_domains, tt::per_element());
+
+  auto& ref_indices = gendb.class_reference_indices;
+
+  // Only the Person field has reference fields.
+  BOOST_TEST(ref_indices.size() == 1);
+  BOOST_TEST_REQUIRE(ref_indices.contains("Person"));
+  BOOST_TEST(ref_indices.at("Person").at(2).at("birth_city") == 0);
+  BOOST_TEST(ref_indices.at("Person").at(2).at("home_city") == 1);
+}
+
+BOOST_AUTO_TEST_CASE(test_domains_and_reference_cache_diamond) {
+  std::stringstream ss(R"""(
+class City
+  name ~ string
+
+class School
+  location ~ City
+
+class Practice
+  location ~ City
+
+class Physician
+  practice ~ Practice
+  school ~ School
+)""");
+  PCleanSchema schema;
+  [[maybe_unused]] bool ok = read_schema(ss, &schema);
+  assert(ok);
+  std::mt19937 prng;
+  GenDB gendb(&prng, schema);
+
+  std::vector<std::string> expected_domains = {"City", "Practice", "City",
+                                               "School", "Physician"};
+  BOOST_TEST(gendb.domains["Physician"] == expected_domains,
+             tt::per_element());
+
+  auto& ref_indices = gendb.class_reference_indices;
+
+  BOOST_TEST(ref_indices.size() == 3);
+
+  // Physician (index 4) has a reference field "practice", which appears
+  // at index 1. Practice has a reference field "location", which appears
+  // at index 0.
+  BOOST_TEST(ref_indices.at("Physician").at(4).at("practice") == 1);
+  BOOST_TEST(ref_indices.at("Physician").at(1).at("location") == 0);
+
+  // Physician (index 4) has a reference field "school", which appears
+  // at index 3. School has a reference field "location", which appears
+  // at index 2.
+  BOOST_TEST(ref_indices.at("Physician").at(4).at("school") == 3);
+  BOOST_TEST(ref_indices.at("Physician").at(3).at("location") == 2);
+
+  BOOST_TEST(ref_indices.at("Practice").at(1).at("location") == 0);
+  BOOST_TEST(ref_indices.at("School").at(1).at("location") == 0);
+}
+
+BOOST_AUTO_TEST_CASE(test_make_relations_for_queryfield) {
+  std::mt19937 prng;
+  GenDB gendb(&prng, schema);
+  T_schema tschema;
+
+  PCleanClass query_class = schema.classes[schema.query.record_class];
+  gendb.make_relations_for_queryfield(schema.query.fields["School"],
+                                              query_class, &tschema);
+
+  BOOST_TEST(tschema.size() == 2);
+  BOOST_TEST(tschema.contains("School"));
+  BOOST_TEST(tschema.contains("Physician::School"));
+  BOOST_TEST(std::get<T_noisy_relation>(tschema["School"]).is_observed);
+  BOOST_TEST(
+      !std::get<T_noisy_relation>(tschema["Physician::School"]).is_observed);
+}
+
+BOOST_AUTO_TEST_CASE(test_make_relations_for_queryfield_only_final_emissions) {
+  std::mt19937 prng;
+  GenDB gendb(&prng, schema, true);
+  T_schema tschema;
+
+  PCleanClass query_class = schema.classes[schema.query.record_class];
+  gendb.make_relations_for_queryfield(schema.query.fields["School"],
+                                              query_class, &tschema);
+  BOOST_TEST(tschema.size() == 1);
+  BOOST_TEST(tschema.contains("School"));
+}
+
+BOOST_AUTO_TEST_CASE(test_make_hirm_schmea) {
+  std::mt19937 prng;
+  GenDB gendb(&prng, schema);
+  T_schema tschema = gendb.make_hirm_schema();
+
+  BOOST_TEST(tschema.contains("School:name"));
+  T_clean_relation cr = std::get<T_clean_relation>(tschema["School:name"]);
+  BOOST_TEST(!cr.is_observed);
+  BOOST_TEST((cr.distribution_spec.distribution == DistributionEnum::bigram));
+  std::vector<std::string> expected_domains = {"School"};
+  BOOST_TEST(cr.domains == expected_domains);
+
+  BOOST_TEST(tschema.contains("School:degree_dist"));
+  T_clean_relation cr2 =
+      std::get<T_clean_relation>(tschema["School:degree_dist"]);
+  BOOST_TEST(
+      (cr2.distribution_spec.distribution == DistributionEnum::categorical));
+  BOOST_TEST(cr2.distribution_spec.distribution_args.contains("k"));
+  BOOST_TEST(cr2.domains == expected_domains);
+
+  BOOST_TEST(tschema.contains("Physician:degree"));
+  T_clean_relation cr3 =
+      std::get<T_clean_relation>(tschema["Physician:degree"]);
+  BOOST_TEST(
+      (cr3.distribution_spec.distribution == DistributionEnum::stringcat));
+  std::vector<std::string> expected_domains2 = {"School", "Physician"};
+  BOOST_TEST(cr3.domains == expected_domains2);
+
+  BOOST_TEST(tschema.contains("Physician:specialty"));
+
+  BOOST_TEST(tschema.contains("City:name"));
+  T_clean_relation cr4 = std::get<T_clean_relation>(tschema["City:name"]);
+  std::vector<std::string> expected_domains3 = {"City"};
+  BOOST_TEST(cr4.domains == expected_domains3);
+
+  BOOST_TEST(tschema.contains("City:state"));
+
+  BOOST_TEST(tschema.contains("Specialty"));
+  T_noisy_relation nr1 = std::get<T_noisy_relation>(tschema["Specialty"]);
+  BOOST_TEST(nr1.is_observed);
+  BOOST_TEST((nr1.emission_spec.emission == EmissionEnum::bigram_string));
+  expected_domains = {"School", "Physician", "Record"};
+  BOOST_TEST(nr1.domains == expected_domains, tt::per_element());
+
+  BOOST_TEST(tschema.contains("School"));
+  T_noisy_relation nr2 = std::get<T_noisy_relation>(tschema["School"]);
+  BOOST_TEST(nr2.is_observed);
+  BOOST_TEST((nr2.emission_spec.emission == EmissionEnum::bigram_string));
+  expected_domains = {"School", "Physician", "Record"};
+  BOOST_TEST(nr2.domains == expected_domains, tt::per_element());
+
+  BOOST_TEST(tschema.contains("Degree"));
+  T_noisy_relation nr3 = std::get<T_noisy_relation>(tschema["Degree"]);
+  BOOST_TEST(nr3.is_observed);
+  BOOST_TEST((nr3.emission_spec.emission == EmissionEnum::bigram_string));
+  expected_domains = {"School", "Physician", "Record"};
+  BOOST_TEST(nr3.domains == expected_domains, tt::per_element());
+
+  BOOST_TEST(tschema.contains("City"));
+  T_noisy_relation nr4 = std::get<T_noisy_relation>(tschema["City"]);
+  BOOST_TEST(nr4.is_observed);
+  BOOST_TEST((nr4.emission_spec.emission == EmissionEnum::bigram_string));
+  expected_domains = {"City", "Practice", "Record"};
+  BOOST_TEST(nr4.domains == expected_domains, tt::per_element());
+
+  BOOST_TEST(tschema.contains("State"));
+  T_noisy_relation nr5 = std::get<T_noisy_relation>(tschema["State"]);
+  BOOST_TEST(nr5.is_observed);
+  BOOST_TEST((nr5.emission_spec.emission == EmissionEnum::bigram_string));
+  expected_domains = {"City", "Practice", "Record"};
+  BOOST_TEST(nr5.domains == expected_domains, tt::per_element());
+
+  BOOST_TEST(tschema.contains("Physician::School"));
+  T_noisy_relation nr6 =
+      std::get<T_noisy_relation>(tschema["Physician::School"]);
+  BOOST_TEST(!nr6.is_observed);
+  expected_domains = {"School", "Physician"};
+  BOOST_TEST(nr6.domains == expected_domains, tt::per_element());
+
+  BOOST_TEST(tschema.contains("Practice::City"));
+  T_noisy_relation nr7 = std::get<T_noisy_relation>(tschema["Practice::City"]);
+  BOOST_TEST(!nr7.is_observed);
+  expected_domains = {"City", "Practice"};
+  BOOST_TEST(nr7.domains == expected_domains, tt::per_element());
+
+  BOOST_TEST(tschema.contains("Practice::State"));
+  T_noisy_relation nr8 = std::get<T_noisy_relation>(tschema["Practice::State"]);
+  BOOST_TEST(!nr8.is_observed);
+  expected_domains = {"City", "Practice"};
+  BOOST_TEST(nr8.domains == expected_domains, tt::per_element());
+
+  auto& ref_indices = gendb.relation_reference_indices;
+
+  // Practice (index 1) has a reference field "city", which appears
+  // at index 0.
+  BOOST_TEST(ref_indices.at("Practice::State").at(1).at("city") == 0);
+
+  // Record (index 2) has a reference field "location", which appears
+  // at index 1 (and refers to Practice). Practice has a reference field
+  // "city", which appears at index 0.
+  BOOST_TEST(ref_indices.at("State").at(2).at("location") == 1);
+  BOOST_TEST(ref_indices.at("State").at(1).at("city") == 0);
+}
+
+BOOST_AUTO_TEST_CASE(test_make_hirm_schema_only_final_emissions) {
+  std::mt19937 prng;
+  GenDB gendb(&prng, schema, true);
+  T_schema tschema = gendb.make_hirm_schema();
+
+  BOOST_TEST(tschema.contains("School:name"));
+  T_clean_relation cr = std::get<T_clean_relation>(tschema["School:name"]);
+  BOOST_TEST(!cr.is_observed);
+  BOOST_TEST((cr.distribution_spec.distribution == DistributionEnum::bigram));
+  std::vector<std::string> expected_domains = {"School"};
+  BOOST_TEST(cr.domains == expected_domains);
+
+  BOOST_TEST(tschema.contains("School:degree_dist"));
+  T_clean_relation cr2 =
+      std::get<T_clean_relation>(tschema["School:degree_dist"]);
+  BOOST_TEST(
+      (cr2.distribution_spec.distribution == DistributionEnum::categorical));
+  BOOST_TEST(cr2.distribution_spec.distribution_args.contains("k"));
+  BOOST_TEST(cr2.domains == expected_domains);
+
+  BOOST_TEST(tschema.contains("Physician:degree"));
+  T_clean_relation cr3 =
+      std::get<T_clean_relation>(tschema["Physician:degree"]);
+  BOOST_TEST(
+      (cr3.distribution_spec.distribution == DistributionEnum::stringcat));
+  std::vector<std::string> expected_domains2 = {"School", "Physician"};
+  BOOST_TEST(cr3.domains == expected_domains2);
+
+  BOOST_TEST(tschema.contains("Physician:specialty"));
+
+  BOOST_TEST(tschema.contains("City:name"));
+  T_clean_relation cr4 = std::get<T_clean_relation>(tschema["City:name"]);
+  std::vector<std::string> expected_domains3 = {"City"};
+  BOOST_TEST(cr4.domains == expected_domains3);
+
+  BOOST_TEST(tschema.contains("City:state"));
+
+  BOOST_TEST(tschema.contains("Specialty"));
+  T_noisy_relation nr1 = std::get<T_noisy_relation>(tschema["Specialty"]);
+  BOOST_TEST(nr1.is_observed);
+  BOOST_TEST((nr1.emission_spec.emission == EmissionEnum::bigram_string));
+  expected_domains = {"School", "Physician", "Record"};
+  BOOST_TEST(nr1.domains == expected_domains, tt::per_element());
+
+  BOOST_TEST(tschema.contains("School"));
+  T_noisy_relation nr2 = std::get<T_noisy_relation>(tschema["School"]);
+  BOOST_TEST(nr2.is_observed);
+  BOOST_TEST((nr2.emission_spec.emission == EmissionEnum::bigram_string));
+  expected_domains = {"School", "Physician", "Record"};
+  BOOST_TEST(nr2.domains == expected_domains, tt::per_element());
+
+  BOOST_TEST(tschema.contains("Degree"));
+  T_noisy_relation nr3 = std::get<T_noisy_relation>(tschema["Degree"]);
+  BOOST_TEST(nr3.is_observed);
+  BOOST_TEST((nr3.emission_spec.emission == EmissionEnum::bigram_string));
+  expected_domains = {"School", "Physician", "Record"};
+  BOOST_TEST(nr3.domains == expected_domains, tt::per_element());
+
+  BOOST_TEST(tschema.contains("City"));
+  T_noisy_relation nr4 = std::get<T_noisy_relation>(tschema["City"]);
+  BOOST_TEST(nr4.is_observed);
+  BOOST_TEST((nr4.emission_spec.emission == EmissionEnum::bigram_string));
+  expected_domains = {"City", "Practice", "Record"};
+  BOOST_TEST(nr4.domains == expected_domains, tt::per_element());
+
+  BOOST_TEST(tschema.contains("State"));
+  T_noisy_relation nr5 = std::get<T_noisy_relation>(tschema["State"]);
+  BOOST_TEST(nr5.is_observed);
+  BOOST_TEST((nr5.emission_spec.emission == EmissionEnum::bigram_string));
+  expected_domains = {"City", "Practice", "Record"};
+  BOOST_TEST(nr5.domains == expected_domains, tt::per_element());
+
+  auto& ref_indices = gendb.relation_reference_indices;
+  BOOST_TEST(ref_indices.at("State").at(2).at("location") == 1);
+  BOOST_TEST(ref_indices.at("State").at(1).at("city") == 0);
+}
+
+BOOST_AUTO_TEST_CASE(test_record_class_is_clean) {
+  std::stringstream ss2(R"""(
+class Record
+  rent ~ real
+
+observe
+  rent as "Rent"
+  from Record
+)""");
+  PCleanSchema schema2;
+  [[maybe_unused]] bool ok = read_schema(ss2, &schema2);
+  assert(ok);
+
+  std::mt19937 prng;
+  GenDB gendb(&prng, schema2, false, true);
+  T_schema tschema = gendb.make_hirm_schema();
+
+  BOOST_TEST(!tschema.contains("Record:rent"));
+  BOOST_TEST(tschema.contains("Rent"));
+
+  T_clean_relation cr = std::get<T_clean_relation>(tschema["Rent"]);
+  BOOST_TEST(cr.is_observed);
+}
+
+BOOST_AUTO_TEST_CASE(test_record_class_is_dirty) {
+  std::stringstream ss2(R"""(
+class Record
+  rent ~ real
+
+observe
+  rent as "Rent"
+  from Record
+)""");
+  PCleanSchema schema2;
+  [[maybe_unused]] bool ok = read_schema(ss2, &schema2);
+  assert(ok);
+
+  std::mt19937 prng;
+  GenDB gendb(&prng, schema2, false, false);
+  T_schema tschema = gendb.make_hirm_schema();
+
+  BOOST_TEST(tschema.contains("Record:rent"));
+  BOOST_TEST(tschema.contains("Rent"));
+
+  T_clean_relation cr = std::get<T_clean_relation>(tschema["Record:rent"]);
+  BOOST_TEST(!cr.is_observed);
+  T_noisy_relation nr = std::get<T_noisy_relation>(tschema["Rent"]);
+  BOOST_TEST(nr.is_observed);
 }
 
 BOOST_AUTO_TEST_SUITE_END()
