+++ cxx/pclean/pclean_lib.hh
@@ -11,10 +11,12 @@
 #include "pclean/schema.hh"
 
 // For each non-missing value in the DataFrame df, create an
-// observation in the returned T_observations.  The column name of the value
+// observation and incorporate it into the GenDB.  The column name of the value
 // is used as the relation name, and each entity in each domain is given
 // its own unique value.
-T_observations translate_observations(const DataFrame& df, GenDB *gendb);
+void incorporate_observations(std::mt19937* prng,
+                              GenDB *gendb,
+                              const DataFrame& df);
 
 // Return a dataframe of num_samples samples from the GenDB.
 DataFrame make_pclean_samples(int num_samples, GenDB *gendb,
